requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bleach==6.1.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-summernote==0.8.20.0
django-tinymce==4.0.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==2.0.1
gunicorn==22.0.0
idna==3.7
ipython==8.26.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
openpyxl==3.1.3
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==1.7.1
requests==2.32.3
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
starkbank-ecdsa==2.2.0
stripe==7.3.0
traitlets==5.13.0
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.12
webencodings==0.5.1
